Anomaly Detector demo: add AnomalyDetectorScenarioLoader
