+++ Kiosk/IntelligentKioskSample.csproj
@@ -205,6 +205,7 @@
     <Compile Include="Views\AnomalyDetector\AnomalyDetectorDemo.xaml.cs">
       <DependentUpon>AnomalyDetectorDemo.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Views\AnomalyDetector\AnomalyDetectorScenarioLoader.cs" />
     <Compile Include="Views\AnomalyDetector\ScenarioInfoControl.xaml.cs">
       <DependentUpon>ScenarioInfoControl.xaml</DependentUpon>
     </Compile>
